Fix broken es2k switchlink_link_test
When switchlink was modified for Linux Networking 2.0, a function
was added that broke switchlink_link_test when it was built for
ES2K. This CL moves that function to a separate file, allowing
the unit test to provide a test dummy.
