+++ switchlink/CMakeLists.txt
@@ -29,6 +29,13 @@ add_library(switchlink_o OBJECT
     $<TARGET_OBJECTS:switchlink_sai_o>
 )
 
+if(ES2K_TARGET)
+    target_sources(switchlink_o PRIVATE
+      switchlink_es2k_utils.c
+      switchlink_es2k_utils.h
+    )
+endif()
+
 target_compile_options(switchlink_o PRIVATE -DHAVE_NLA_BITFIELD32)
 
 if(WITH_OVSP4RT)
