+++ switchlink/switchlink_es2k_utils.h
@@ -1,6 +1,5 @@
 /*
- * Copyright 2013-present Barefoot Networks, Inc.
- * Copyright 2022-2023 Intel Corporation.
+ * Copyright 2024 Intel Corporation.
  * SPDX-License-Identifier: Apache 2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,17 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#ifndef __SWITCHLINK_ES2K_UTILS__
+#define __SWITCHLINK_ES2K_UTILS__
 
-#ifndef __SWITCHLINK_UTILS_H__
-#define __SWITCHLINK_UTILS_H__
-
-#include <netinet/in.h>
 #include <stdbool.h>
-#include <stdint.h>
 
-#include "switchutils/switch_utils.h"
+bool switchlink_validate_driver(const char* ifname);
 
-uint32_t ipv4_prefix_len_to_mask(uint32_t prefix_len);
-struct in6_addr ipv6_prefix_len_to_mask(uint32_t prefix_len);
-
-#endif /* __SWITCHLINK_UTILS_H__ */
+#endif  // __SWITCHLINK_ES2K_UTILS__
