+++ switchlink/switchlink_link.c
@@ -17,16 +17,12 @@
  */
 
 #include <fcntl.h>
-#include <linux/errno.h>
-#include <linux/ethtool.h>
 #include <linux/if.h>
 #include <linux/if_bridge.h>
-#include <linux/sockios.h>
 #include <linux/version.h>
 #include <netlink/attr.h>
 #include <netlink/msg.h>
 #include <netlink/netlink.h>
-#include <sys/ioctl.h>
 #include <unistd.h>
 
 #include "switchlink.h"
@@ -34,7 +30,11 @@
 #include "switchlink_handlers.h"
 #include "switchlink_int.h"
 #include "switchlink_link_types.h"
-#include "switchlink_utils.h"
+#include "switchutils/switch_log.h"
+
+#ifdef ES2K_TARGET
+#include "switchlink_es2k_utils.h"
+#endif
 
 #if defined(ES2K_TARGET)
 // ES2K creates netdevs from idpf driver/SR-IOVs.
@@ -92,50 +92,6 @@ static switchlink_link_type_t get_link_type(const char* info_kind) {
 
   return link_type;
 }
-
-#if defined(ES2K_TARGET)
-/*
- * Routine Description:
- *    Check if the interface driver is valid for our use case
- *
- * Arguments:
- *    [in] ifname - Interface name
- *
- * Return Values:
- *    boolean
- */
-bool validate_driver_name(char* ifname) {
-  struct ethtool_drvinfo drv = {0};
-  char drvname[32] = {0};
-  struct ifreq ifr = {0};
-  int fd, r = 0;
-
-  fd = socket(AF_INET, SOCK_DGRAM, 0);
-  if (fd < 0) {
-    return false;
-  }
-
-  drv.cmd = ETHTOOL_GDRVINFO;
-  strncpy(ifr.ifr_name, ifname, sizeof(ifr.ifr_name));
-  ifr.ifr_data = (void*)&drv;
-
-  r = ioctl(fd, SIOCETHTOOL, &ifr);
-  if (r) {
-    goto end;
-  }
-
-  strncpy(drvname, drv.driver, sizeof(drvname));
-
-  if (!memcmp(drvname, "openvswitch", strlen(drvname)) ||
-      !memcmp(drvname, "idpf", strlen(drvname))) {
-    close(fd);
-    return true;
-  }
-end:
-  close(fd);
-  return false;
-}
-#endif
 
 /*
  * Routine Description:
@@ -407,7 +363,7 @@ void switchlink_process_link_msg(const struct nlmsghdr* nlmsg, int msgtype) {
         }
 
 #if defined(ES2K_TARGET)
-        if (!validate_driver_name(attrs.ifname)) {
+        if (!switchlink_validate_driver(attrs.ifname)) {
           krnlmon_log_info(
               "Ignoring interface: %s which is not created"
               " by openvswitch or idpf driver",
