+++ switchlink/switchlink_link_test.cc
@@ -14,6 +14,9 @@ extern "C" {
 #include "switchlink_handlers.h"
 #include "switchlink_int.h"
 #include "switchlink_link_types.h"
+#ifdef ES2K_TARGET
+#include "switchlink_es2k_utils.h"
+#endif
 }
 
 #define IPV4_ADDR(a, b, c, d) (((a) << 24) | ((b) << 16) | ((c) << 8) | (d))
@@ -79,6 +82,9 @@ void switchlink_delete_lag(uint32_t ifindex) {}
 void switchlink_create_lag_member(
     switchlink_db_lag_member_info_t* lag_member_info) {}
 void switchlink_delete_lag_member(uint32_t ifindex) {}
+
+bool switchlink_validate_driver(const char* ifname) { return true; }
+
 #endif
 
 void switchlink_create_interface(switchlink_db_interface_info_t* intf) {
